feat: rename spark to distributed
