+++ edsnlp/processing/wrapper.py
@@ -33,7 +33,6 @@ def pipe(
         - `n_jobs>1` corresponds to `parallel_pipe` with `n_jobs` parallel workers
         - `n_jobs=-1` corresponds to `parallel_pipe` with maximun number of workers
         - `n_jobs=-2` corresponds to `parallel_pipe` with maximun number of workers -1
-
     additional_spans : Union[List[str], str], by default "discarded"
         A name (or list of names) of SpanGroup on which to apply the pipe too:
         SpanGroup are available as `doc.spans[spangroup_name]` and can be generated
@@ -84,25 +83,12 @@ def pipe(
             """  # noqa W291
         )
 
-    from .spark import pipe as spark_pipe
+    from .distributed import pipe as distributed_pipe
 
-    if module == DataFrameModules.PYSPARK:
-
-        return spark_pipe(
-            note=note,
-            nlp=nlp,
-            additional_spans=additional_spans,
-            extensions=extensions,
-            **kwargs,
-        )
-
-    if module == DataFrameModules.KOALAS:
-        note_spark = note.to_spark()
-        note_nlp = spark_pipe(
-            note=note_spark,
-            nlp=nlp,
-            additional_spans=additional_spans,
-            extensions=extensions,
-            **kwargs,
-        )
-        return note_nlp.to_koalas()
+    return distributed_pipe(
+        note=note,
+        nlp=nlp,
+        additional_spans=additional_spans,
+        extensions=extensions,
+        **kwargs,
+    )
